Feat: #108 UpdateElderInfoRepository 인터페이스 패턴 리팩토링
- UpdateElderInfoRepository를 인터페이스로 분리
- UpdateElderInfoRepositoryImpl 구현체 생성
- 노인 정보 수정/삭제 기능 인터페이스 적용
